+++ meilisearch-http/src/routes/mod.rs
@@ -6,7 +6,7 @@ pub mod health;
 pub mod index;
 pub mod key;
 pub mod search;
-pub mod setting;
+pub mod settings;
 pub mod stats;
 
 #[derive(Serialize)]
